Remove macro CASSERT
CASSERT does not seem to make sense to me as continuable errors are
not advised to be used in general, and an assertion is meant to be a
static check beyond which program execution should not continue.
Therefore continuing from a failed assertion seems to be a bad idea.

All call locations to macro CASSERT have been updated to use CL:ASSERT
instead.